common: build warning fixes

+++ src/src/common.cpp
@@ -125,7 +125,10 @@ uint8_t BindingUID[6] = {0, 1, 2, 3, 4, 5}; // Special binding UID values
         uint8_t UID[6];
         esp_err_t WiFiErr = esp_read_mac(UID, ESP_MAC_WIFI_STA);
     #elif PLATFORM_STM32
-        uint8_t UID[6] = {HAL_GetUIDw0(), HAL_GetUIDw0() >> 8, HAL_GetUIDw1(), HAL_GetUIDw1() >> 8, HAL_GetUIDw2(), HAL_GetUIDw2() >> 8};
+        uint8_t UID[6] = {
+            (uint8_t)HAL_GetUIDw0(), (uint8_t)(HAL_GetUIDw0() >> 8),
+            (uint8_t)HAL_GetUIDw1(), (uint8_t)(HAL_GetUIDw1() >> 8),
+            (uint8_t)HAL_GetUIDw2(), (uint8_t)(HAL_GetUIDw2() >> 8)};
     #else
         uint8_t UID[6] = {0};
     #endif
